BUG FIXES: VARS-517, VARS-17

+++ vars-annotation/src/main/java/vars/annotation/ui/actions/UpdateNewRefNumAction.java
@@ -95,7 +95,7 @@ public class UpdateNewRefNumAction extends ActionAdapter implements IVideoArchiv
 
                     // Get the largest ref number and add 1
                     if (!intValues.isEmpty()) {
-                       refNum = ((Integer) Collections.max(intValues)).intValue();
+                       refNum = Collections.max(intValues);
                        refNum++;
                     }
